new feature for orchestrate
Added a new feature in the library to enable the orchestrate API.
It's not enabled by default.

Will only be compiled in when this feature is enabled.
