+++ src/protocols/multi_party_ecdsa/gg_2020/mod.rs
@@ -15,6 +15,8 @@
 */
 
 pub mod blame;
+#[cfg(any(feature = "orchestrategg20", test))]
+pub mod orchestrate;
 pub mod party_i;
 #[cfg(test)]
 mod test;
